+++ IDay.cs
@@ -3,6 +3,6 @@ namespace adventofcode2021
     interface IDay
     {
         public void Test(string input);
-        public int Run(string input);
+        public decimal Run(string input);
     }
 }
